[REFACTOR] Migrate Full Content View to Independent Composable
  - Modify import statement with new path applied to DetailedContentState
