+++ app/src/main/java/com/doyoonkim/knutice/domain/CrawlFullContentImpl.kt
@@ -1,7 +1,7 @@
 package com.doyoonkim.knutice.domain
 
 import com.doyoonkim.knutice.data.NoticeLocalRepository
-import com.doyoonkim.knutice.viewModel.DetailedContentState
+import com.doyoonkim.knutice.model.DetailedContentState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
